schemas moved to trace_api.py
